Comparator renamed
